Fix Dockerfile
There was a possibility to improperly build an image that would contain
docs from mismatched TSLint version. This could result in error caused
by the engine being unable to find issue descriptions.

This Dockerfile eliminates the issue and adds a few clean up steps to
bring the image size a little down.